Make travis build run just unit tests, move integration tests into a proper folder
