+++ .travis.yml
@@ -10,7 +10,7 @@ node_js:
 install: npm install
 
 script:
-  - npm test
+  - npm run test:unit
 
 notifications:
   email: false
